Stop iframe-ing Kancolle page. Instead open a new browser window for it.
This is because it's impossible, or at least terribly hard to manipulate
an element in an iframe. Same origin policy will be applied...
Using the server as a proxy for the src of the iframe would make it at
least possible, but it involves postMessage() and implementing another
proxy service, which looks quite complicated. Also, Polymer.dart
elements cannot be extracted for non-Dart clients -- Selenium is not an
exception.

Rather, I can have 2 separate windows, 1 for Kancolle and 1 for KCAA UI.
This way Kancolle page's elments can be directly controlled via
Selenium.